refactor: reorganize media-related services and update package structure for clarity
- Move and rename service classes to align with new package organization
- Update imports and references to reflect new locations
- Rename VariantGeneratorService to ImageVariantGenerator for specificity
- Consolidate upload and storage logic under clearer namespaces
