+++ file/src/main/java/pt/estga/file/services/MediaContentService.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Service;
-import pt.estga.file.services.storage.FileStorageService;
+import pt.estga.file.storage.FileStorageService;
 
 import java.io.InputStream;
