+++ file/src/main/java/pt/estga/file/services/MediaMetadataService.java
@@ -1,4 +1,4 @@
-package pt.estga.file.services.metadata;
+package pt.estga.file.services;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
