+++ file/src/main/java/pt/estga/file/services/upload/MediaValidationService.java
@@ -1,4 +1,4 @@
-package pt.estga.file.services;
+package pt.estga.file.services.upload;
 
 import lombok.extern.slf4j.Slf4j;
 import org.apache.tika.Tika;
